backend/services/paper-storage-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv>=1.0.1
pydantic==2.5.3
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
minio==7.2.3
python-multipart==0.0.6
pdfplumber==0.11.0
pypdf2==3.0.1
httpx==0.27.0
aiohttp==3.9.1

